Don't add Liquid to `this` when not using exports
Liquid.js leaks global variables when defined inside a closure that has
an implicit `this` of the global window object, e.g.,

(function() {
  // Liquid source code here
})();

My reasoning behind this change is that if Liquid is required directly
in the browser, it will be added to the global window object anyway (so
this does not change any behavior there). However, if Liquid is defined
inside a closure, it will not leak the `Liquid` variable.